Add HTTP GET utility
The utility consists out of multi-state updatable downloader objects that are able to further process or delegate the data if desired.
They all are constructed with the same builder which will proceed into an implementation-specific state once specific settings have been made.

The provided classes use implementation hiding via the builder pattern and interfaces. The implementation is based on org.apache.hc.
The latter has been added as dependency to the pom.xml