Extract AnnotationDiscovery interface to AnnotationUtil temporarily
This is the first part of the AnnotationDiscovery rewrite.
Created a new implementation of the extracted interface.
It should provide the same range of functionality, but with the difference that subsequent filter calls will merge those filters.
It has slightly extended api and is more precise in its actions.
A lot of elements are marked with fixme where further design choices have to be made.