Deprecate AnnotationDiscovery, Update AnnotationUtil, Add AnnotationMap
Deprecate the old annotation system, regardless whether the approach of the AnnotationUtilImpl or the AnnotationMap will be continued.

Create an alternative way to access annotations, the AnnotationMap
It will scan the whole class and all of its methods at once and buffer the results. It provides various methods to limit the results and is capable to create result views which limit the complete results to a small part, allowing easier access.

Pull up some members of the AnnotationUtilImpl to the AnnotationUtil
This has the purpose of providing legacy support. These methods have been added to the AnnotationDiscovery as well.